Moved the persistent storage functions to a separate file so that they can be used by other components. Re-organized the globus download steps function so that sign in tokens aren't required when setting the path of a saved endpoint. Also simplified the flow of the function to follow based on user goals rather than page state.
