+++ frontend/src/components/Cart/Summary.tsx
@@ -11,7 +11,7 @@ import { RawSearchResult, RawSearchResults } from '../Search/types';
 import { UserCart } from './types';
 import { GlobusTaskItem } from '../Globus/types';
 import GlobusStateKeys, { globusTaskItems } from '../Globus/recoil/atom';
-import { saveSessionValue } from '../../api';
+import { addPersistVar, setPersistVal } from '../../common/persistData';
 
 const styles: CSSinJS = {
   headerContainer: { display: 'flex', justifyContent: 'center' },
@@ -35,6 +35,7 @@ const Summary: React.FC<React.PropsWithChildren<Props>> = ({ userCart }) => {
   const [taskItems, setTaskItems] = useRecoilState<GlobusTaskItem[]>(
     globusTaskItems
   );
+  addPersistVar(GlobusStateKeys.globusTaskItems, [], setTaskItems);
 
   let numFiles = 0;
   let totalDataSize = '0';
@@ -53,8 +54,7 @@ const Summary: React.FC<React.PropsWithChildren<Props>> = ({ userCart }) => {
   }
 
   const clearAllTasks = async (): Promise<void> => {
-    setTaskItems([]);
-    await saveSessionValue(GlobusStateKeys.globusTaskItems, []);
+    await setPersistVal(GlobusStateKeys.globusTaskItems, [], true);
   };
 
   return (
